requirements.txt
Django>=3.0.5,<3.6
djangorestframework>=3.11.0,<3.12.0
psycopg2>=2.8.5
gunicorn>=20,<21
django-cors-headers>=2.0,<=3.5

scrapy-djangoitem==1.1.1

flake8>=3.0,<=3.9

Scrapy>=2,<3
scrapy-user-agents >=0.1.1,<=2